New debug and logging logic.

+++ bin/freesoundoftheday.lua
@@ -10,47 +10,50 @@ local API_ENDPOINT = "https://freesound.org/apiv2/"
 local API_TOKEN    = "7GpLkaFEHHF4m2zPzhmKyfZEgCP0G1l12VqVGKg7"
 local DIRECTORY    = "/home/we/dust/audio/freesoundoftheday/"
 
-local debug = true
+local DEBUG        = false
+local LOGGING      = true
+
+if DEBUG then DIRECTORY = "/tmp/freesoundoftheday/" end
 
 --------------- network connections -----------------
 
 function get_frontpage(path)
    local content = nil
    if (path ~= nil) then
-      if debug then print("🗄 Getting frontpage from file.") end
+      log("🗄 Getting frontpage from file.")
       local fd = io.open(path, 'r')
       content = fd:read '*all'
       fd:close()
-      if debug then print("🗜 Read "..#content.."B frontpage from file.") end
+      log("🗜 Read "..#content.."B frontpage from file.")
    else
-      if debug then print("📡 Getting frontpage from network.") end
+      log("📡 Getting frontpage from network.")
       local response = assert(io.popen("curl -s https://freesound.org"))
       content = response:read '*all'
       response:close()
-      if debug then print("🗜 Read "..#content.."B frontpage from network.") end
+      log("🗜 Read "..#content.."B frontpage from network.")
    end
    return content
 end
 
 function get_featured_sound_id(frontpage)
-   if debug then print("🔐 Looking for sound id.") end
+   log("🔐 Looking for sound id.")
    -- assumes it's the first player on the page. Sad.
    -- local pattern = 'class="sample_player_small" id="(%d+)"'
    local pattern = 'Random sound of the day.*data.sound.id="(%d+)"'
    local featured_sound_id = string.match(frontpage, pattern)
-   if debug then print("🔑 Found sound id "..featured_sound_id) end
+   log("🔑 Found sound id "..featured_sound_id)
    return featured_sound_id
 end
 
 function get_sound_metadata(sound_id)
-   if debug then print("📡 Getting metadata for sound id "..sound_id) end
+   log("📡 Getting metadata for sound id "..sound_id)
    local metadata = query_api("sounds", sound_id)
    local parsed   = parse_metadata(metadata)
    return parsed
 end
 
 function get_sound_preview(metadata)
-   if debug then print("💿 Downloading preview for "..metadata['name']) end
+   log("💿 Downloading preview for "..metadata['name'])
    local filename = extract_preview_filename(metadata)
    local complete_path = "/tmp/"..filename
    download_file(metadata['previews']['preview-hq-mp3'], complete_path)
@@ -60,7 +63,7 @@ end
 --------------- lower level network connections -------
 
 function download_file(url, filename)
-   if debug then print("📡 Downloading file from "..url) end
+   log("📡 Downloading file from "..url)
    -- io.popen("curl -s "..url.. " -o "..DIRECTORY..filename)
    io.popen("curl -s "..url.." -o "..filename)
 end
@@ -69,11 +72,12 @@ function query_api(resource, param)
    -- Just takes a single parameter rather than a table.
    local query_url = API_ENDPOINT..resource.."/"..param.."/"
    local auth_url = query_url.."?token="..API_TOKEN
-   if debug then print("🛂 "..auth_url) end
+   log("🛂 "..auth_url)
    local response = assert(io.popen("curl -s "..auth_url..";sync"))
    local content = response:read '*all'
    response:close()
-   -- if debug then print("📇 Read metadata "..content) end
+   -- log("📇 Read metadata "..content)
+   log("📇 Read "..#content.. "B of metadata ")
    return content
 end
 
@@ -91,8 +95,8 @@ end
 
 function sound_already_on_disk(directory, metadata)
    local filename_would_be = create_richer_preview_filename(metadata)
-   if debug then print("🗃 Checking for file "..filename_would_be
-                       .." in directory "..directory) end
+   log("🗃 Checking for file "..filename_would_be
+                       .." in directory "..directory)
    local files = {}
    for filename in lfs.dir(directory) do
       if (filename == filename_would_be) then return true end
@@ -105,10 +109,10 @@ function purge_all_except(directory, metadata)
    for filename in lfs.dir(directory) do
       if filename ~= "." and filename ~= ".." then
          if filename ~= keeper then
-            if debug then print("🗑 Binning "..filename) end
-            io.popen("rm "..directory.."/"..'"'..filename..'"')
+            log("🗑 Binning "..filename)
+            if not DEBUG then io.popen("rm "..directory.."/"..'"'..filename..'"') end
          else
-            if debug then print("♻ Would keep "..filename) end
+            log("♻ Would keep "..filename)
          end
       end
    end
@@ -116,13 +120,17 @@ end
 
 function render_to_wav(fromfile, tofile)
    -- Because OAuth2 is hard lol.
-   if debug then print("🔊 Rendering file "..fromfile) end
+   log("🔊 Rendering file "..fromfile)
    local wavfilename = tofile:gsub('.mp3$', '.wav')
-   if debug then print("🔊 Rendering into "..wavfilename) end
+   log("🔊 Rendering into "..wavfilename)
    io.popen("nice ffmpeg -n -loglevel warning -i "..fromfile.." '"..wavfilename.."'")
 end
 
 ----------------- utilities -----------------
+
+function log(message)
+   if LOGGING then print(message) end
+end
 
 function parse_metadata(metadata)
    return lunajson.decode(metadata)
@@ -137,24 +145,25 @@ end
 
 function extract_preview_filename(metadata)
    local filename = metadata['previews']['preview-hq-mp3']:gsub('^.*/', '')
-   if debug then print("🗜 preview filename is "..filename) end
+   log("🗜 preview filename is "..filename)
    return filename
 end
 
 ---------- main -----------
 
 function main()
+   if DEBUG then print("🔍 Debugging on") end
    -- local fp = get_frontpage('frontpage.html')
    local fp = get_frontpage()
    local sound_id = get_featured_sound_id(fp)
    local sound_of_the_day = get_sound_metadata(sound_id)
-   print("Today's random sound of the day is "
+   print("👂 Today's random sound of the day is "
          ..sound_of_the_day['name'].." by "..sound_of_the_day['username'])
 
    if sound_already_on_disk(DIRECTORY, sound_of_the_day) then
-      if debug then print("⏹ Already got "..sound_of_the_day['name']) end
+      log("⏹ Already got "..sound_of_the_day['name'])
    else
-      if debug then print("▶ Getting random sound of the day.") end
+      log("▶ Getting random sound of the day.")
       local tmp_preview_filename = get_sound_preview(sound_of_the_day)
       local rendered_preview_filename = DIRECTORY
          ..create_richer_preview_filename(sound_of_the_day)
